requirements.txt
Flask~=1.1.2
numpy~=1.18.5
matplotlib~=3.2.2
tensorflow~=2.3.0
Pillow~=7.2.0